Move `gio::compile_resources` to its own crate
Closes #574
