+++ .github/workflows/CI.yml
@@ -116,6 +116,8 @@ jobs:
       - name: Check two levels glib dependent
         run: cargo check
         working-directory: tests/two-levels-glib-dependent
+      - name: "glib-build-tools: build"
+        run: cargo build --manifest-path glib-build-tools/Cargo.toml
       # examples
       - name: "examples"
         run: cargo build --manifest-path examples/Cargo.toml --bins --examples --all-features
