+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "gdk-pixbuf/sys",
   "gio",
   "gio/sys",
+  "glib-build-tools",
   "glib",
   "glib/gobject-sys",
   "glib/sys",
