+++ examples/Cargo.toml
@@ -20,8 +20,8 @@ path = "../gio"
 version = "0.11"
 optional = true
 
-[build-dependencies.gio]
-path = "../gio"
+[build-dependencies.glib-build-tools]
+path = "../glib-build-tools"
 
 
 [[bin]]
