+++ examples/build.rs
@@ -1,5 +1,5 @@
 fn main() {
-    gio::compile_resources(
+    glib_build_tools::compile_resources(
         "resources",
         "resources/resources.gresource.xml",
         "compiled.gresource",
