+++ glib-build-tools/COPYRIGHT
@@ -1,0 +1,1 @@
+../COPYRIGHT
