+++ glib-build-tools/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "glib-build-tools"
+license = "MIT"
+homepage = "https://gtk-rs.org/"
+authors = ["The gtk-rs Project Developers"]
+keywords = ["glib", "gio", "gtk-rs", "gnome", "build-dependencies"]
+readme = "README.md"
+documentation = "https://gtk-rs.org/gtk-rs-core/stable/latest/docs/glib-build-tools"
+version = "0.1.0"
+description = "Rust bindings for the Gio library, build script utils crate"
+repository = "https://github.com/gtk-rs/gtk-rs-core"
+edition = "2021"
+rust-version = "1.57"
+
+[dependencies]
+gio = { path = "../gio", optional = true }
+
+[features]
+# We only depend on gio so that we may link it in our documentation
+dox = ["gio"]
