+++ glib-build-tools/README.md
@@ -1,0 +1,16 @@
+# GIO build helpers
+
+Crate containing helpers for building GIO-based applications.
+
+## Minimum supported Rust version
+
+Currently, the minimum supported Rust version is `1.57.0`.
+
+## Documentation
+
+* [Stable API](https://gtk-rs.org/gtk-rs-core/stable/latest/docs/glib-build-tools)
+* [Development API](https://gtk-rs.org/gtk-rs-core/git/latest/docs/glib-build-tools)
+
+### See Also
+
+* [gio](http://crates.io/crates/gio)
